Small refatoring
* Extracted the seconds_ago function into a separate module
* Extracted the release tasks into a separate module
