+++ lib/sanbase/release_tasks/migrate.ex
@@ -1,0 +1,39 @@
+defmodule Sanbase.ReleaseTasks.Migrate do
+  @start_apps [
+    :crypto,
+    :ssl,
+    :postgrex,
+    :ecto
+  ]
+
+  def repos, do: Application.get_env(:sanbase, :ecto_repos, [])
+
+  def run do
+    IO.puts "Starting dependencies.."
+    # Start apps necessary for executing migrations
+    Enum.each(@start_apps, &Application.ensure_all_started/1)
+
+    # Start the Repo(s) for myapp
+    IO.puts "Starting repos.."
+    Enum.each(repos(), &(&1.start_link(pool_size: 1)))
+
+    # Run migrations
+    Enum.each(repos(), &run_migrations_for/1)
+  end
+
+  defp run_migrations_for(repo) do
+    app = Keyword.get(repo.config, :otp_app)
+    IO.puts "Running migrations for #{app}"
+    Ecto.Migrator.run(repo, migrations_path(repo), :up, all: true)
+  end
+
+  defp migrations_path(repo), do: priv_path_for(repo, "migrations")
+
+  defp priv_path_for(repo, filename) do
+    app = Keyword.get(repo.config, :otp_app)
+    repo_underscore = repo |> Module.split |> List.last |> Macro.underscore
+    Path.join([priv_dir(app), repo_underscore, filename])
+  end
+
+  defp priv_dir(app), do: "#{:code.priv_dir(app)}"
+end
