+++ rel/commands/migrate.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-$RELEASE_ROOT_DIR/bin/sanbase command Elixir.Sanbase.ReleaseTasks migrate
+$RELEASE_ROOT_DIR/bin/sanbase command Elixir.Sanbase.ReleaseTasks run
